Initial Frontend Test Suite Move
